requirements.txt
Jinja2==3.1.2
Markdown==3.4.3
PyYAML==6.0
pygments==2.15.1
python-markdown-math